+++ CHANGELOG.md
@@ -12,6 +12,8 @@ In Development
 - `pyrepo mkgithub`: Create custom labels used by `.github/dependabot.yml`
   template
 - `pyrepo mkgithub`: Properly sanitize keywords used as topics
+- Use [`ghtoken`](https://github.com/jwodder/ghtoken) for looking up GitHub
+  tokens instead of storing them in the pyrepo config file
 - Templates:
     - Add 'A' to enabled flake8 checks in `tox.ini`
     - Update coverage paths config in `tox.ini` for Coverage v7